refactor models, continued work on addpurchase logic and view

+++ Modules/Core/Models/Period.swift
@@ -11,9 +11,9 @@ public struct Period {
 	public let id: String
 	public let periodStart: Date
 	public let periodEnd: Date
-	public let amountAvailable: Decimal
+	public let amountAvailable: Int
 	
-	public init(id: String, periodStart: Date, periodEnd: Date, amountAvailable: Decimal) {
+	public init(id: String, periodStart: Date, periodEnd: Date, amountAvailable: Int) {
 		self.id = id
 		self.periodStart = periodStart
 		self.periodEnd = periodEnd
